Remove missing unit checks in scheduler. (#104)
There cannot be missing dependents because a unit only declares
its dependencies. This makes it impossible to have missing units
after the missing starting units are removed before being added
to the initial stack.

This PR removes the logic to check for missing units and adds more
missing dependencies to the tests to make sure the scheduler works
with missing dependencies.