Give more weight to more recent scores. (#117)
Previously, if an user tries a new exercise multiple times in the first few
days, the weights of their scores will be the same. This will make it
hard to get a passing score if the user makes fast progress.

This change adds a step to adjust the weight of the scores based on the index
of the trial in the list.